+++ lib/src/api/column.dart
@@ -1,20 +1,19 @@
-import 'dart:convert';
-import 'package:http/http.dart' as http;
-
-import 'package:apiraiser/src/helpers/state.dart';
-import 'package:apiraiser/src/helpers/headers.dart';
+import 'package:apiraiser/src/api/rest.dart';
 import 'package:apiraiser/src/models/column_info.dart';
 import 'package:apiraiser/src/models/api_result.dart';
+import 'package:apiraiser/src/models/rest_params.dart';
 
 /// Column APIs
 class Column {
   /// Get predefined columns
   Future<APIResult> getPredefinedColumns() async {
     try {
-      var res = await http.get(
-          Uri.parse('${State.endPoint}/API/v1/GetPredefinedColumns'),
-          headers: Headers.getHeaders());
-      return APIResult.fromJson(json.decode(res.body));
+      var res = await Rest.get(
+        RestParams(
+          '/API/v1/GetPredefinedColumns',
+        ),
+      );
+      return APIResult.fromJson(res);
     } catch (e) {
       return APIResult(message: e.toString(), success: false);
     }
@@ -23,10 +22,12 @@ class Column {
   /// Get columns
   Future<APIResult> get(String table) async {
     try {
-      var res = await http.get(
-          Uri.parse('${State.endPoint}/API/v1/$table/Columns'),
-          headers: Headers.getHeaders());
-      return APIResult.fromJson(json.decode(res.body));
+      var res = await Rest.get(
+        RestParams(
+          '/API/v1/$table/Columns',
+        ),
+      );
+      return APIResult.fromJson(res);
     } catch (e) {
       return APIResult(message: e.toString(), success: false);
     }
@@ -36,12 +37,11 @@ class Column {
   Future<APIResult> insert(String table, ColumnInfo columnInfo) async {
     Map<String, dynamic> data = columnInfo.toJson(columnInfo);
     try {
-      var res = await http.post(
-        Uri.parse('${State.endPoint}/API/v1/$table/Column'),
-        headers: Headers.getHeaders(),
-        body: jsonEncode(data),
-      );
-      return APIResult.fromJson(json.decode(res.body));
+      var res = await Rest.post(RestParams(
+        '/API/v1/$table/Column',
+        data: data,
+      ));
+      return APIResult.fromJson(res);
     } catch (e) {
       return APIResult(message: e.toString(), success: false);
     }
@@ -50,11 +50,10 @@ class Column {
   /// Remove column
   Future<APIResult> delete(String table, String column) async {
     try {
-      var res = await http.delete(
-        Uri.parse('${State.endPoint}/API/v1/$table/Column/$column'),
-        headers: Headers.getHeaders(),
+      var res = await Rest.delete(
+        RestParams('/API/v1/$table/Column/$column'),
       );
-      return APIResult.fromJson(json.decode(res.body));
+      return APIResult.fromJson(res);
     } catch (e) {
       return APIResult(message: e.toString(), success: false);
     }