+++ lib/src/api/importexport.dart
@@ -1,10 +1,10 @@
 import 'dart:typed_data';
 import 'package:apiraiser/apiraiser.dart';
-import 'package:apiraiser/src/helpers/headers.dart' as headers;
-import 'package:dio/dio.dart';
-import 'package:http/http.dart' as http;
-
+import 'package:apiraiser/src/api/rest.dart';
 import 'package:apiraiser/src/helpers/state.dart';
+
+import 'package:apiraiser/src/models/rest_params.dart';
+import 'package:dio/dio.dart';
 
 /// ImportExport APIs
 class ImportExport {
@@ -12,12 +12,12 @@ class ImportExport {
   Future<Uint8List?> exportExcel(
       String table, List<QuerySearchItem> conditions) async {
     try {
-      var res = await http.post(
-        Uri.parse('${State.endPoint}/API/v1/ImportExport/ExportExcel/$table'),
-        headers: headers.Headers.getHeaders(),
-        body: QuerySearchItem.toJsonList(conditions),
+      var res = await Rest.post(
+        RestParams('/API/v1/ImportExport/ExportExcel/$table',
+            data: QuerySearchItem.toJsonList(conditions),
+            responseType: ResponseType.stream),
       );
-      return res.bodyBytes;
+      return res;
     } catch (e) {
       rethrow;
     }